great refactoring
