--- src/api/index.ts
@@ -1,2 +1,0 @@
-export { login } from './login'
-export { register } from './register'
