--- src/api/login.ts
@@ -1,10 +1,0 @@
-import { handleFetch } from "./services";
-import { IUserLogin } from "./types";
-
-export const login = async (event: React.FormEvent, userData: IUserLogin) => {
-  event.preventDefault();
-
-  const result = await handleFetch('login', 'POST', userData);
-  console.log('result: ', result)
-  return result;
-}
