+++ src/types.ts
@@ -17,4 +17,7 @@ export interface IResponse {
   }
 }
 
-export type Method = 'POST' | 'GET'
+export enum Method {
+  POST = 'POST',
+  GET = 'GET'
+}
