+++ vite.config.ts
@@ -12,9 +12,9 @@ export default defineConfig({
       '@constants': path.resolve(__dirname, 'src/constants'),
       '@helpers': path.resolve(__dirname, 'src/helpers'),
       '@i18n': path.resolve(__dirname, 'src/i18n'),
-      '@api': path.resolve(__dirname, 'src/api'),
       '@hooks': path.resolve(__dirname, 'src/hooks'),
       '@utils': path.resolve(__dirname, 'src/utils'),
+      '@services': path.resolve(__dirname, 'src/services'),
     }
   }
 })
